finalized the treatment
