+++ src/components/AnimalRoutes.js
@@ -1,5 +1,6 @@
 import React from "react"
 import { Route } from "react-router-dom"
+import { AddTreatment } from "./animals/AddTreatMent"
 import {Animal} from "./animals/Animal"
 import AnimalForm from "./animals/AnimalForm"
 import { AnimalListComponent } from "./animals/AnimalList"
@@ -15,6 +16,9 @@ export default () => {
             </Route>
             <Route path="/animals/new">
                 <AnimalForm />
+            </Route>
+            <Route path="/addtreatment/:animalId(\d+)">
+                <AddTreatment />
             </Route>
         </>
     )
